#4 extract UiSettings from GoogleMapController
