+++ maps-google/src/commonMain/kotlin/dev/icerock/moko/maps/google/UiSettings.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright 2019 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package dev.icerock.moko.maps.google
+
+data class UiSettings(
+    val compassEnabled: Boolean = false,
+    val myLocationEnabled: Boolean = false,
+    val myLocationButtonEnabled: Boolean = false,
+    val indoorLevelPickerEnabled: Boolean = false,
+    val scrollGesturesEnabled: Boolean = true,
+    val zoomGesturesEnabled: Boolean = true,
+    val tiltGesturesEnabled: Boolean = true,
+    val rotateGesturesEnabled: Boolean = true,
+    val scrollGesturesDuringRotateOrZoomEnabled: Boolean = true
+)
